Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/fetch/http-cache/cc-request-expected.txt


PASS HTTP cache doesn't use aged but fresh response when request contains Cache-Control: max-age=0 
FAIL HTTP cache doesn't use aged but fresh response when request contains Cache-Control: max-age=1 assert_equals: Response 2 comes from cache expected 2 but got 1
FAIL HTTP cache doesn't use fresh response with Age header when request contains Cache-Control: max-age that is greater than remaining freshness assert_equals: Response 2 comes from cache expected 2 but got 1
PASS HTTP cache does use aged stale response when request contains Cache-Control: max-stale that permits its use 
PASS HTTP cache does reuse stale response with Age header when request contains Cache-Control: max-stale that permits its use 
FAIL HTTP cache doesn't reuse fresh response when request contains Cache-Control: min-fresh that wants it fresher assert_equals: Response 2 comes from cache expected 2 but got 1
FAIL HTTP cache doesn't reuse fresh response with Age header when request contains Cache-Control: min-fresh that wants it fresher assert_equals: Response 2 comes from cache expected 2 but got 1
PASS HTTP cache doesn't reuse fresh response when request contains Cache-Control: no-cache 
PASS HTTP cache validates fresh response with Last-Modified when request contains Cache-Control: no-cache 
PASS HTTP cache validates fresh response with ETag when request contains Cache-Control: no-cache 
FAIL HTTP cache doesn't reuse fresh response when request contains Cache-Control: no-store assert_equals: Response 2 comes from cache expected 2 but got 1
FAIL HTTP cache generates 504 status code when nothing is in cache and request contains Cache-Control: only-if-cached promise_test: Unhandled rejection with value: object "TypeError: can’t load from network"
